Pan-Card-Tampering-Detection-master/requirements.txt
Flask==2.2.3
Pillow==9.4.0
scikit-image==0.20.0
gunicorn
